Check errors if renaming a primary type without refactoring
